backend/requirements.txt
fastapi==0.63.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
celery~=4.4.7
uvicorn~=0.13.4
pyvips>=2.1.14
pydantic~=1.8.1
aiofiles~=0.6.0
mysql-connector-python
SQLAlchemy~=1.4.2
starlette~=0.13.6
pydantic[email]
shortuuid~=1.0.1
requests~=2.25.1
opencv-python>=4.5.3.56
Shapely~=1.7.1
numpy~=1.20.3
imutils~=0.5.4
